+++ docs/source/changelog.rst
@@ -14,6 +14,25 @@ New Features
 Other Changes
 -------------
 
+* None.
+
+Bugfixes
+--------
+
+* None.
+
+======================
+  0.1.11 (2024-09-25)
+======================
+
+New Features
+------------
+
+* None.
+
+Other Changes
+-------------
+
 * Updates to documentation.
 * Use ``__slots__`` to improve resource use and immutability.
 * Use ``__all__`` to manage exports from module
@@ -21,7 +40,7 @@ Other Changes
 Bugfixes
 --------
 
-* ``range`` raises ``ValueError`` if argument value for step is ``0``.
+* ``range()`` raises ``ValueError`` if argument value for step is ``0``.
 
 ======================
   0.1.10 (2024-09-23)